Revert "Revert "MLOPS-320: Update MS2DeepScore to match the queries against db spectra""
