+++ notebooks/ms2deepscore/GNPS-2-SPECTRA.mgf
@@ -1,0 +1,146 @@
+BEGIN IONS
+PEPMASS=353.232
+CHARGE=1
+MSLEVEL=2
+SOURCE_INSTRUMENT=LC-ESI-Maxis II HD Q-TOF Bruker
+FILENAME=JnJ_10uL_B3_GB3_01_2587.mzXML
+SEQ=*..*
+IONMODE=positive
+ORGANISM=GNPS-COLLECTIONS-MISC
+NAME=Eicosanoids_8-iso-15kPGF2?_C20H32O5 M+H
+PI=Dorrestein/Wu
+DATACOLLECTOR=lfnothias
+SMILES=n/a
+INCHI=n/a
+INCHIAUX=N/A
+PUBMED=n/a
+SUBMITUSER=mwang87
+LIBRARYQUALITY=1
+SPECTRUMID=CCMSLIB00001058176
+SCANS=596
+79.054993	176.0
+81.071831	541.0
+91.05558	1471.0
+93.066986	169.0
+97.067764	368.0
+100.526466	310.0
+102.323349	180.0
+105.035217	522.0
+105.072456	611.0
+107.05098	152.0
+107.08213	181.0
+109.027458	172.0
+115.054199	558.0
+117.071159	594.0
+118.079468	314.0
+119.091988	540.0
+120.090439	328.0
+121.067108	323.0
+128.064468	418.0
+129.07074	1249.0
+130.071472	206.0
+131.084152	1879.0
+133.065414	614.0
+134.074036	224.0
+141.069687	480.0
+142.076782	677.0
+143.058212	98.0
+143.0867	1017.0
+144.182053	69.0
+145.105743	469.0
+146.106125	263.0
+149.095551	100.0
+152.683319	81.0
+154.075775	175.0
+155.089249	1139.0
+156.084747	241.0
+157.103882	1045.0
+157.128403	129.0
+157.453781	173.0
+159.082367	172.0
+159.114914	392.0
+163.107849	114.0
+165.071884	80.0
+168.096436	130.0
+169.104828	100.0
+171.118118	610.0
+172.121017	233.0
+173.098862	377.0
+173.137177	189.0
+173.363235	153.0
+174.099808	207.0
+175.07991	204.0
+175.112991	341.0
+179.110931	137.0
+179.150391	112.0
+181.076447	256.0
+182.026276	144.0
+183.117706	531.0
+184.122284	382.0
+185.096268	539.0
+191.136963	97.0
+193.09552	164.0
+193.121216	201.0
+199.128693	259.0
+201.101089	182.0
+201.128174	767.0
+203.108093	191.0
+204.11499	102.0
+204.313797	61.0
+207.140228	492.0
+210.108047	97.0
+211.098846	91.0
+217.166153	349.0
+219.14624	214.0
+225.123505	129.0
+229.119797	257.0
+237.161697	128.0
+253.155197	194.0
+255.105392	139.0
+255.169312	148.0
+281.168182	273.0
+281.192688	330.0
+299.207123	972.0
+317.213196	937.0
+318.214539	246.0
+335.213806	262.0
+335.381714	182.0
+336.231049	152.0
+END IONS
+
+BEGIN IONS
+PEPMASS=377.232
+CHARGE=1
+MSLEVEL=2
+SOURCE_INSTRUMENT=LC-ESI-Maxis II HD Q-TOF Bruker
+FILENAME=JnJ_10uL_B7_GB7_01_2592.mzXML
+SEQ=*..*
+IONMODE=positive
+ORGANISM=GNPS-COLLECTIONS-MISC
+NAME=Eicosanoids_ResolvinD1_C22H32O5 M+H
+PI=Dorrestein/Wu
+DATACOLLECTOR=lfnothias
+SMILES=n/a
+INCHI=n/a
+INCHIAUX=N/A
+PUBMED=n/a
+SUBMITUSER=mwang87
+LIBRARYQUALITY=1
+SPECTRUMID=CCMSLIB00001058177
+SCANS=738
+153.792694	151.0
+175.03894	165.0
+241.088303	158.0
+309.44809	286.0
+321.101929	257.0
+375.889862	74.0
+376.199432	90.0
+376.761658	207.0
+377.230042	8316.0
+377.739319	121.0
+377.870697	70.0
+378.237823	1605.0
+378.529327	59.0
+379.616089	52.0
+411.208832	130.0
+END IONS
