+++ tests/ms2deepscore/test_ms2deepscore_integration.py
@@ -3,17 +3,12 @@ import pytest
 
 @pytest.mark.internet_connection
 @pytest.mark.skip(reason="Requires a valid token")
-def test_predict_similarity_of_pair(mgf_path_of_pair, ms2deepscore_client):
-    result = ms2deepscore_client.predict_similarity_of_pair(
-        mgf_path_of_pair,
+def test_match_spectra_from_path_with_2_spectra(mgf_path_of_pair, ms2deepscore_client):
+    result = ms2deepscore_client.match_spectra_from_path(
+        mgf_path_of_pair, 3, ["compound_name"]
     )
     assert result
+    assert set(result[0].columns) == {"score", "compound_name"}
 
 
-@pytest.mark.internet_connection
-@pytest.mark.skip(reason="Requires a valid token")
-def test_match_spectra_from_path(mgf_path_of_equal_pair, ms2deepscore_client):
-    result = ms2deepscore_client.predict_similarity_of_pair(mgf_path_of_equal_pair)
-
-    assert result
-    assert result["Tanimoto Score"] == 1
+# TODO: add more tests
