+++ tests/test_endpoint.py
@@ -1,22 +1,11 @@
-import pandas as pd
 import pytest
 import requests
-
-from omigami import Spec2Vec
-from omigami.spec2vec import InvalidCredentials
-
-
-@pytest.fixture()
-def prediction_endpoints():
-    _client = Spec2Vec("")
-    return {
-        "positive": _client._PREDICT_ENDPOINT_BASE.format(ion_mode="positive"),
-        "negative": _client._PREDICT_ENDPOINT_BASE.format(ion_mode="negative"),
-    }
+import pandas as pd
+from omigami.endpoint import Endpoint, InvalidCredentials
 
 
 def test_build_payload(mgf_generator):
-    client = Spec2Vec("token")
+    client = Endpoint("token")
 
     payload = client._build_payload((mgf_generator), {"n_best_spectra": 10})
 
@@ -26,8 +15,8 @@ def test_build_payload(mgf_generator):
 
 
 @pytest.mark.internet_connection
-def test_unauthorized_request(prediction_endpoints):
-    client = Spec2Vec("bad_token")
+def test_unauthorized_request(spec2vec_prediction_endpoints):
+    client = Endpoint("bad_token")
     small_payload = {
         "data": {
             "ndarray": {
@@ -43,11 +32,11 @@ def test_unauthorized_request(prediction_endpoints):
     }
 
     with pytest.raises(InvalidCredentials):
-        client._send_request(small_payload, prediction_endpoints["positive"])
+        client._send_request(small_payload, spec2vec_prediction_endpoints["positive"])
 
 
 def test_format_results(sample_response):
-    client = Spec2Vec("token")
+    client = Endpoint("token")
     requests.Response()
 
     results = client._format_results(sample_response)
@@ -57,44 +46,45 @@ def test_format_results(sample_response):
     assert all(results[0]["score"] > 0)
 
 
+def test_validate_parameters():
+    parameters = Endpoint._build_parameters(2, ["smiles"])
+    assert {"n_best_spectra", "include_metadata"} == set(parameters.keys())
+
+    with pytest.raises(ValueError, match="batman"):
+        _ = Endpoint._build_parameters(2, ["batman"])
+
+    with pytest.raises(ValueError, match="must be an integer"):
+        _ = Endpoint._build_parameters("robin", ["smiles"])
+
+
 def test_validate_input():
     model_input = {
         "peaks_json": "[[80.060677, 157.0], [337.508301, 230.0]]",
         "Precursor_MZ": "153.233",
     }
+    endpoint = Endpoint("token")
     # first validates if the input is correct then we test for errors
-    Spec2Vec._validate_input([model_input])
+    endpoint._validate_input([model_input])
 
     with pytest.raises(TypeError, match="Spectrum data must be a dictionary."):
-        Spec2Vec._validate_input(["not_a_dict"])
+        endpoint._validate_input(["not_a_dict"])
 
     with pytest.raises(KeyError, match="mandatory keys"):
-        Spec2Vec._validate_input([{"Precursor_MZ": "1", "peaks_JASON": "[not a list]"}])
+        endpoint._validate_input([{"Precursor_MZ": "1", "peaks_JASON": "[not a list]"}])
 
     with pytest.raises(
         ValueError, match="peaks_json needs to be a valid python string representation"
     ):
-        Spec2Vec._validate_input([{"Precursor_MZ": "1", "peaks_json": "[not a list]"}])
+        endpoint._validate_input([{"Precursor_MZ": "1", "peaks_json": "[not a list]"}])
 
     with pytest.raises(
         ValueError,
         match="peaks_json needs to be a valid python string representation",
     ):
-        Spec2Vec._validate_input([{"Precursor_MZ": "1", "peaks_json": 10}])
+        endpoint._validate_input([{"Precursor_MZ": "1", "peaks_json": 10}])
 
     with pytest.raises(
         ValueError,
         match="Precursor_MZ needs to be a string representation of a float",
     ):
-        Spec2Vec._validate_input([{"Precursor_MZ": "float", "peaks_json": [10]}])
-
-
-def test_validate_parameters():
-    parameters = Spec2Vec._build_parameters(2.0, ["smiles"])
-    assert {"n_best_spectra", "include_metadata"} == set(parameters.keys())
-
-    with pytest.raises(ValueError, match="batman"):
-        _ = Spec2Vec._build_parameters(2, ["batman"])
-
-    with pytest.raises(ValueError, match="must be an integer"):
-        _ = Spec2Vec._build_parameters("robin", ["smiles"])
+        endpoint._validate_input([{"Precursor_MZ": "float", "peaks_json": [10]}])
